src/requirements.txt
fastapi[all]
uvicorn[standard]
opencensus-ext-azure